+++ widgets/calphacontrolwidget.cpp
@@ -32,7 +32,7 @@ void CAlphaControlWidget::updateAlphaControlWidget(int alpha)
     m_alphaSlider->blockSignals(true);
     m_alphaSlider->setValue(alpha);
     m_alphaSlider->blockSignals(false);
-    m_alphaLabel->setText(QString("%1%").arg((int)(alpha * 100 / 255)));
+    m_alphaLabel->setText(QString("%1%").arg(int(alpha * 100 / 255)));
 }
 
 int CAlphaControlWidget::getCurrentAlphaValue()
@@ -93,7 +93,7 @@ void CAlphaControlWidget::initUI()
 void CAlphaControlWidget::initConnection()
 {
     connect(m_alphaSlider, &DSlider::valueChanged, this, [ = ](int value) {
-        m_alphaLabel->setText(QString("%1%").arg((int)value * 100 / 255));
+        m_alphaLabel->setText(QString("%1%").arg(value * 100 / 255));
         emit signalAlphaChanged(value);
     });
